% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_master.R
\docType{class}
\name{app_master}
\alias{app_master}
\title{R6 class for storing information across the app.}
\format{
A \code{\link{R6Class}} generator object
}
\description{
app_master
}
\examples{
\dontrun{
app_m <- app_master$new()
}
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{app_master$new()}}
\item \href{#method-preload_master}{\code{app_master$preload_master()}}
\item \href{#method-row}{\code{app_master$row()}}
\item \href{#method-dataset_names}{\code{app_master$dataset_names()}}
\item \href{#method-data_by_name}{\code{app_master$data_by_name()}}
\item \href{#method-data_by_index}{\code{app_master$data_by_index()}}
\item \href{#method-clone}{\code{app_master$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-preload_master"></a>}}
\if{latex}{\out{\hypertarget{method-preload_master}{}}}
\subsection{Method \code{preload_master()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$preload_master(master_data)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-row"></a>}}
\if{latex}{\out{\hypertarget{method-row}{}}}
\subsection{Method \code{row()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$row()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dataset_names"></a>}}
\if{latex}{\out{\hypertarget{method-dataset_names}{}}}
\subsection{Method \code{dataset_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$dataset_names()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_by_name"></a>}}
\if{latex}{\out{\hypertarget{method-data_by_name}{}}}
\subsection{Method \code{data_by_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$data_by_name(dataset_name)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-data_by_index"></a>}}
\if{latex}{\out{\hypertarget{method-data_by_index}{}}}
\subsection{Method \code{data_by_index()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$data_by_index(index)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{app_master$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
