% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{create_app_r}
\alias{create_app_r}
\title{Creates the app.R using the templates defined in inst/cheetah/*}
\usage{
create_app_r(env = "default")
}
\arguments{
\item{env}{(optional) defauls to "default". Switch this to production , dev , test as need be}
}
\description{
Based on the configuration defined in config.yml the app.R code is created
}
