% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{create_dash_components}
\alias{create_dash_components}
\title{Creates the dash_components.R using the templates defined in inst/cheetah/*}
\usage{
create_dash_components(env = "default")
}
\arguments{
\item{env}{(optional) defauls to "default". Switch this to production , dev , test as need be}
}
\description{
Based on the configuration defined in config.yml the dash_components.R code is created
}
