% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{create_files}
\alias{create_files}
\title{Create the files based on iputs params}
\usage{
create_files(dots)
}
\arguments{
\item{dots}{the inputs params as defined in newproject.dcf}
}
\description{
Create the files based on iputs params
}
