% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_spring.R
\name{create_new_project}
\alias{create_new_project}
\title{CLI option to create a new Shiny Spring Project}
\usage{
create_new_project(
  dashboard_template = "bs4_dash",
  app_type = "minimal",
  config_file = "config.yml"
)
}
\arguments{
\item{dashboard_template}{(optional) defaults to "bs4_dash". Options are shiny_dashboard_plus , argon_dash}

\item{app_type}{(optional) default to "basic". Options are}

\item{config_file}{(optional)}
}
\value{
TRUE
}
\description{
CLI option to create a new Shiny Spring Project
}
