% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_data_upload.R
\name{create_row}
\alias{create_row}
\title{creates the standardized row for app_master mdata}
\usage{
create_row(srnum, filename, ds_name, ds)
}
\description{
creates the standardized row for app_master mdata
}
