% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{get_params}
\alias{get_params}
\title{get parameters from the config.yml file}
\usage{
get_params(env = "default", path = "config.yml")
}
\arguments{
\item{env}{(optional) defauls to "default". Switch this to production , dev , test as need be}

\item{path}{(optional) default to config.yml in the root project directory}
}
\value{
parameters
}
\description{
Loads the config.yml file using the config package
}
