% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_gen.R
\name{load_config_template}
\alias{load_config_template}
\title{Load the config.yml template available in inst/cheetah}
\usage{
load_config_template()
}
\value{
the txt file for config.yml
}
\description{
Reads to file inst/cheetah/config.yml returns the file
}
